Everything broken again it seems

+++ src/user/doomgeneric/i_system.h
@@ -32,7 +32,7 @@ void I_Init (void);
 // Called by startup code
 // to get the ammount of memory to malloc
 // for the zone management.
-byte*	I_ZoneBase (int *size);
+byte*	I_ZoneBase (size_t *size);
 
 boolean I_ConsoleStdout(void);
